Sfx_PlayFromObjectLimited: int sfxId param recovers full-width FindObjectChannel arg (#58)
Target passes sfxId via plain mr to Sfx_FindObjectChannel (u32 param) with no
clrlwi; a u16 param forced a spurious narrowing. Recover the original int param
type (DLL externs already declare it int). Eliminates the call-1 clrlwi.

+++ include/main/audio/sfx.h
@@ -9,7 +9,7 @@ void Sfx_ClearLoopedObjectSounds(void);
 void Sfx_UpdateLoopedObjectSounds(void);
 void Sfx_SetObjectSoundsPaused(s32 paused);
 void Sfx_PlayFromObject(u32 obj, u16 sfxId);
-u32 Sfx_PlayFromObjectLimited(u32 obj, u16 sfxId, int limit);
+u32 Sfx_PlayFromObjectLimited(u32 obj, int sfxId, int limit);
 void Sfx_PlayAtPositionFromObject(f32 x, f32 y, f32 z, u32 obj, u16 sfxId);
 void Sfx_PlayFromObjectChannel(u32 obj, u32 channel, u16 sfxId);
 void Sfx_StopFromObject(u32 obj, u32 sfxId);
